requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
supabase==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
Pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
face-recognition==1.3.0
dlib==19.24.2
pandas==2.2.2
aiofiles==23.2.1
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1
jinja2==3.1.4
aiosmtplib==3.0.1
python-dateutil==2.9.0
